# doc/swagger.yaml
{
  "swagger": "2.0",
  "info": {
    "description": "This is a simple user login API",
    "title": "[REST] User Login API",
    "version": "1.0.0"
  },
  "basePath": "/api/v1",
  "paths": {
    "/users": {
      "post": {
        "description": "create User",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "users"
        ],
        "summary": "Create User",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUserStruct"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            },
            "examples": {
              "application/json": {
                "id": 0,
                "username": "",
                "email": "",
                "created_at": "0001-01-01T00:00:00Z"
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "put": {
        "description": "update User",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "users"
        ],
        "summary": "Update User",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateUserRequest"
            }
          },
          {
            "type": "integer",
            "format": "int64",
            "description": "ID of the resource",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            },
            "examples": {
              "application/json": {
                "id": 0,
                "username": "",
                "email": "",
                "created_at": "0001-01-01T00:00:00Z"
              }
            }
          }
        }
      },
      "delete": {
        "description": "delete User",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "users"
        ],
        "summary": "Delete User",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "description": "ID of the resource",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "definitions": {
    "CreateUserStruct": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "email": {
          "description": "User's email address",
          "type": "string",
          "example": "john@example.com"
        },
        "first_name": {
          "description": "First name of the user",
          "type": "string",
          "example": "John"
        },
        "id": {
          "description": "Unique identifier for the user",
          "type": "integer",
          "format": "int64",
          "example": "1"
        },
        "last_name": {
          "description": "Last name of the user",
          "type": "string",
          "example": "Doe"
        },
        "username": {
          "description": "Username for login",
          "type": "string",
          "example": "john_doe"
        }
      }
    },
    "UpdateUserRequest": {
      "type": "object",
      "properties": {
        "update_email": {
          "description": "Update the email of the user",
          "type": "string",
          "example": "johnny@example.com"
        },
        "update_username": {
          "description": "Update the username of the user",
          "type": "string",
          "example": "johnny_bravo"
        }
      }
    },
    "UserResponse": {
      "type": "object",
      "properties": {
        "email": {
          "description": "User's email address",
          "type": "string",
          "example": "john@example.com"
        },
        "id": {
          "description": "Unique identifier for the user",
          "type": "integer",
          "format": "int64",
          "example": "1"
        },
        "username": {
          "description": "Username for login",
          "type": "string",
          "example": "john_doe"
        }
      }
    }
  }
}